fix: handle imported classes differently to avoid upvalue limit
Reduce the number of upvalues referenced within a closure by
changing the way that `new Class(...)` is transpiled into Lua.

The old way added a file-local variable for each class that was
imported from a module, and `new Class(...)` directly upvalued
to the file-local variable. Unfortunately, if a closure created
instances of more than 60 classes, then the WoW client limit of
60 upvalues is reached and the addon fails to load.

The new way adds a `__imports` table that contains all of the
imported symbols, and `new Class(...)` is transpiled into
`__imports.Class(...)`. This uses only a single upvalue for the
import table instead of an upvalue for each class constructor.

Sort all of the additions to the `__imports` table ahead of all of
the local aliasing of the imported symbols to make it more visually
pleasing.