fix: only define __imports table if it is needed
Change how to determine if the `__imports` table is needed. Instead
of only checking whether non-global modules are imported from, just
re-use the existing check of whether or not an imported symbol from
a non-global module is used before defining the `__imports` table.

This change avoids a possible situation where an `__imports` table
is defined in the Lua output, but no key-value pairs are added to
the table.