fix: handle imported classes differently to avoid upvalue limit (#81)
* fix: handle imported classes differently to avoid upvalue limit

Reduce the number of upvalues referenced within a closure by
changing the way that `new Class(...)` is transpiled into Lua.

The old way added a file-local variable for each class that was
imported from a module, and `new Class(...)` directly upvalued
to the file-local variable. Unfortunately, if a closure created
instances of more than 60 classes, then the WoW client limit of
60 upvalues is reached and the addon fails to load.

The new way adds a `__imports` table that contains all of the
imported symbols, and `new Class(...)` is transpiled into
`__imports.Class(...)`. This uses only a single upvalue for the
import table instead of an upvalue for each class constructor.

Sort all of the additions to the `__imports` table ahead of all of
the local aliasing of the imported symbols to make it more visually
pleasing.

* fix: only define __imports table if it is needed

Change how to determine if the `__imports` table is needed. Instead
of only checking whether non-global modules are imported from, just
re-use the existing check of whether or not an imported symbol from
a non-global module is used before defining the `__imports` table.

This change avoids a possible situation where an `__imports` table
is defined in the Lua output, but no key-value pairs are added to
the table.