fix: transpile TypeScript for statement into Lua while loop (#100)
* fix: transpile TypeScript for statement into Lua while loop

The way that TypeScript `for` loops were transpile into Lua assumed
that all `for` loops are of the form:

    for (let i = 0; i <= 10; i += 2) { ... }

which gets transpiled into:

    for i = 0, 10, 2 do ... end

This requires that all for loops have this very strict form where
the condition MUST be of the form `i <= limit` where the comparator
MUST be `<=`, and the step expression MUST be of the form `i += step`
where the assignment operator MUST be `+=`.

This is very limiting.

For correctness, transpile `for` loops into Lua `while` loops as:

    do
        local i = 0
        while i <= 10 do
            ...
            i = i + 1
        end
    end

This still works correctly if the `...` statements include a
`break` or `continue` statement. More generally, `for` loops in the
general form:

    for (initExpr; conditionExpr; finalExpr) { ...  }

become:

    do
        initExpr
        while conditionExpr do
            ...
            finalExpr
        end
    end

and the outer `do ... end` block scope is not emitted if `initExpr`
isn't a local variable declaration.

* fix: translate simple numeric for loops into Lua equivalent

For certain Typescript `for` loops that meet some heuristic criteria,
emit a Lua numeric `for` loop.

Adjust some of the `for` loop unit tests to cover more cases.